Rename GeneratePlugin to GenerateCommand since it's really a multicommand now and not really a plugin itself. Also move it to system for now.
